requirements.txt
# Core runtime
Flask>=2.3,<3.0
waitress>=2.1,<3.0
requests>=2.31,<3.0

# UI
PySide6>=6.5,<7.0

# Speech / audio
gTTS>=2.3,<3.0
pygame>=2.5,<3.0

# Google Sheets (optional)
gspread>=5.10,<6.0
google-auth>=2.22,<3.0

# Convenience
python-dotenv>=1.0,<2.0

# Optional fuzzy search (client will work without it)
rapidfuzz>=3.2,<4.0

# Tooling
black>=23.7,<24.0
ruff>=0.1,<0.3
mypy>=1.6,<2.0
